+++ src/XalanSourceTree/XalanSourceTreeComment.hpp
@@ -632,7 +632,7 @@ public:
 	appendSiblingNode(XalanSourceTreeText*	theSibling);
 
 	void
-	setIndex(unsigned int	theIndex)
+	setIndex(unsigned long	theIndex)
 	{
 		m_index = theIndex;
 	}
@@ -664,7 +664,7 @@ private:
 
 	XalanNode*						m_nextSibling;
 
-	unsigned int					m_index;
+	unsigned long					m_index;
 
 	static const XalanDOMString&	s_nameString;
 };
