+++ src/XalanSourceTree/XalanSourceTreeText.hpp
@@ -651,7 +651,7 @@ public:
 	appendSiblingNode(XalanSourceTreeText*	theSibling);
 
 	void
-	setIndex(unsigned int	theIndex)
+	setIndex(unsigned long	theIndex)
 	{
 		m_index = theIndex;
 	}
@@ -681,7 +681,7 @@ private:
 
 	XalanNode*						m_nextSibling;
 
-	unsigned int					m_index;
+	unsigned long					m_index;
 
 	static const XalanDOMString&	s_nameString;
 };
